Quickfix: don't crash on non-UTF8 filenames
The program started crashing all of a sudden with the message:

```
rbdb: Cannot decode byte '\xed': Data.Text.Internal.Encoding: Invalid UTF-8 stream
```

It took a while to find the issue — laziness bit me hard because parsing
filenames seemed fine, even the length of the parsed map was correct,
and the crash happened after that. But there is no text decoding after
that! So the thing here must be that the map is keeping its values not
fully evaluated…

The specific bytes causing the error are: `eda0bd edb2a9`. They are
supposed to represent this emoji `💩` (I hate emojis!), which is
`U+1F4A9`, or `f0 9f 92 a9` in UTF-8, or `d8 3d dc a9` in UTF-16.

I don't know how those bytes were generated. The quick fix for now is to
use the lenient decoder to replace invalid bytes with `�`.