Change `Entry.filePath` type from `FilePath` to lazy `Text`
It's technically a more generic type, but `FilePath` is a `String`,
which is not recommended.

Since we may not need to output all filepaths (however we do check
whether they are in `/podcasts`), it may be better to store them as lazy
texts so that only the required paths are evaluated. I'm not sure about
how this works.